pyproject.toml
[project]
name = "nexign-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "faker>=40.1.2",
    "fastapi>=0.128.0",
    "graphviz>=0.21",
    "levenshtein>=0.27.3",
    "matplotlib>=3.10.8",
    "mlflow>=1.27.0",
    "pandas>=3.0.0",
    "peft>=0.18.1",
    "pillow>=12.1.0",
    "protobuf==3.20.3",
    "python-multipart>=0.0.22",
    "qwen-vl-utils>=0.0.14",
    "reportlab>=4.4.9",
    "svglib>=1.6.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
    "transformers>=5.0.0",
    "unicorn>=2.1.4",
    "uvicorn>=0.40.0",
]
